feat: :zap: migrated from cjs to module
