+++ package.json
@@ -2,6 +2,7 @@
   "name": "maven-publish-action",
   "version": "1.0.0",
   "description": "GitHub Action for automatically publishing to Maven central",
+	"type": "module",
   "author": {
     "name": "Wasiq Bhamla",
     "email": "wasbhamla2005@gmail.com",
